refactor(tailer): move the optional-field readers out of parser.go
CodeScene reported core/pkg/tailer/parser.go declining 8.41 -> 8.16 after
#1799 added OptInt and OptDurationFromMillis to it.

They are generic JSON field accessors, not part of ParsedEvent's vocabulary,
and parser.go is already one of the package's largest files — so they get their
own optional.go rather than a suppression. The "an absent number is not zero"
argument moves with them and is still stated once for both.

No behaviour change; the replay goldens stay byte-identical.

Refs #1799
